refactor rearrangement of corner
