+++ src/parser/circle_parser.cpp
@@ -36,78 +36,6 @@ std::vector<cv::Vec3f> detect_circles(cv::Mat img) {
     return detected_circles;
 }
 
-float findArea(cv::Point2f a, cv::Point2f b, cv::Point2f c) {
-    float side_a = sqrt(pow(b.x - c.x, 2) + pow(b.y - c.y, 2));
-    float side_b = sqrt(pow(a.x - c.x, 2) + pow(a.y - c.y, 2));
-    float side_c = sqrt(pow(a.x - b.x, 2) + pow(a.y - b.y, 2));
-
-    if (side_a < 0 || side_b < 0 || side_c < 0 || (side_a + side_b <= side_c) || side_a + side_c <= side_b ||
-        side_b + side_c <= side_a) {
-        throw std::invalid_argument("Invalid sides\n");
-    }
-    float s = (side_a + side_b + side_c) / 2;
-    return sqrt(s * (s - side_a) * (s - side_b) * (s - side_c));
-}
-
-static bool compare_area(const std::pair<float, cv::Point2f>& a, const std::pair<float, cv::Point2f>& b) {
-    return a.first < b.first;
-}
-
-int identify_corner(std::vector<cv::Vec3f>& detected_circles, std::vector<cv::Point2f>& corner_points,
-                    DetectedBarcode& corner_barcode) {
-    corner_points.resize(4);
-
-    int found_mask = 0x00;
-
-    corner_points[BOTTOM_RIGHT] = center_of_box(corner_barcode.bounding_box);
-    found_mask |= (1 << BOTTOM_RIGHT);
-
-    std::pair<float, cv::Point2f> max_distance = { 0, cv::Point2f(0, 0) };
-
-    for (const auto& c : detected_circles) {
-        float x = c[0] - corner_points[BOTTOM_RIGHT].x;
-        float y = c[1] - corner_points[BOTTOM_RIGHT].y;
-        float distance = sqrt(pow(x, 2) + pow(y, 2));
-        if (distance > max_distance.first) {
-            max_distance = { distance, cv::Point2f(c[0], c[1]) };
-        }
-    }
-
-    corner_points[TOP_LEFT] = max_distance.second;
-    found_mask |= (1 << TOP_LEFT);
-
-    std::vector<std::pair<float, cv::Point2f>> right_corner_points;
-
-    for (const auto& circle : detected_circles) {
-        if (circle[0] == max_distance.second.x && circle[1] == max_distance.second.y)
-            continue;
-
-        float angle_corner =
-            angle(corner_points[TOP_LEFT], cv::Point2f(circle[0], circle[1]), corner_points[BOTTOM_RIGHT]);
-        angle_corner = angle_corner < 0 ? angle_corner + M_PI : angle_corner;
-        angle_corner = abs(angle_corner - M_PI / 2);
-        right_corner_points.push_back({ angle_corner, cv::Point2f(circle[0], circle[1]) });
-    }
-
-    std::sort(right_corner_points.begin(), right_corner_points.end(), compare_area);
-
-    if (right_corner_points.size() < 2) {
-        return found_mask;
-    }
-
-    corner_points[TOP_RIGHT] = right_corner_points[0].second;
-    found_mask |= (1 << TOP_RIGHT);
-
-    corner_points[BOTTOM_LEFT] = right_corner_points[1].second;
-    found_mask |= (1 << BOTTOM_LEFT);
-
-    if (corner_points[TOP_RIGHT].x < corner_points[BOTTOM_LEFT].x) {
-        std::swap(corner_points[TOP_RIGHT], corner_points[BOTTOM_LEFT]);
-    }
-
-    return found_mask;
-}
-
 std::optional<cv::Mat> main_circle(cv::Mat img,
 #ifdef DEBUG
                                    cv::Mat debug_img,
@@ -161,7 +89,13 @@ std::optional<cv::Mat> main_circle(cv::Mat img,
 #endif
 
     std::vector<cv::Point2f> corner_points;
-    auto mask = identify_corner(detected_circles, corner_points, corner_barcode);
+    std::vector<cv::Point2f> circle_pos;
+
+    for (const auto& c : detected_circles) {
+        circle_pos.push_back(cv::Point2f(c[0], c[1]));
+    }
+
+    auto mask = found_other_point(circle_pos, corner_points, corner_barcode);
 
 #ifdef DEBUG
     for (int i = 0; i < 4; ++i) {
